% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembler-package.R
\docType{package}
\name{assembler-package}
\alias{assembler}
\alias{assembler-package}
\title{assembler: Assembles R Packages and Projects}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BWAM/assembler}
  \item Report bugs at \url{https://github.com/BWAM/assembler/issues}
}

}
\author{
\strong{Maintainer}: Zachary Smith \email{zachary.smith2@dec.ny.ogv}

}
\keyword{internal}
